+++ src/module.ts
@@ -108,6 +108,8 @@ const module: NuxtModule<ModuleOptions> = defineNuxtModule<ModuleOptions>({
         { name: 'refField', from: resolver.resolve('./runtime/fields/declare') },
         { name: 'refFields', from: resolver.resolve('./runtime/fields/declare') },
         { name: 'mergeFields', from: resolver.resolve('./runtime/fields/declare') },
+        { name: 'defineStep', from: resolver.resolve('./runtime/steps/declare') },
+        { name: 'refSteps', from: resolver.resolve('./runtime/steps/create') },
         { name: 'addRule', from: resolver.resolve('./runtime/fields/register') },
         { name: 'addRules', from: resolver.resolve('./runtime/fields/register') },
         { name: 'addSchema', from: resolver.resolve('./runtime/fields/register') },